Moves Algo metadata to structures, add tests.
